complete appointment
