+++ lib/screens/appointments/components/appointment_type.dart
@@ -1,0 +1,199 @@
+import 'package:flutter/material.dart';
+import 'package:petcarepal/screens/appointments/components/create_medication.dart';
+import 'package:petcarepal/screens/appointments/components/create_vaccination.dart';
+import 'package:petcarepal/screens/appointments/components/create_meal.dart';
+import 'package:petcarepal/screens/appointments/components/create_measurement.dart';
+
+class ChooseAppointmentTypeScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(
+          'Chọn loại lịch',
+          style: TextStyle(
+            fontSize: 18,
+            fontWeight: FontWeight.w700,
+          ),
+        ),
+        centerTitle: true,
+        leading: IconButton(
+          icon: Icon(Icons.arrow_back),
+          onPressed: () {
+            Navigator.pop(context);
+          },
+        ),
+        iconTheme: IconThemeData(color: Color(0xFF4552CB)),
+      ),
+      body: Column(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          GestureDetector(
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => CreateMedicationAppointmentScreen(),
+                ),
+              );
+            },
+            child: Container(
+              width: 500,
+              height: 130,
+              margin: EdgeInsets.all(20),
+              decoration: BoxDecoration(
+                  border: Border.all(color: Colors.white, width: 1),
+                  borderRadius: BorderRadius.circular(10),
+                  image: DecorationImage(
+                      image: AssetImage('assets/images/thuoc.png'),
+                      fit: BoxFit.cover)),
+              child: Container(
+                margin: EdgeInsets.only(left: 30, top: 10),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      'Thuốc',
+                      style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 30,
+                          fontWeight: FontWeight.bold),
+                    ),
+                    Text(
+                      'Quản lí sức khỏe\n loại thuốc chức năng',
+                      style: TextStyle(color: Colors.white, fontSize: 20),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+          GestureDetector(
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => CreateVaccinationAppointmentScreen(),
+                ),
+              );
+            },
+            child: Container(
+              width: 500,
+              height: 130,
+              margin: EdgeInsets.all(20),
+              decoration: BoxDecoration(
+                  border: Border.all(color: Colors.white, width: 1),
+                  borderRadius: BorderRadius.circular(10),
+                  image: DecorationImage(
+                      image: AssetImage('assets/images/tiemchung.png'),
+                      fit: BoxFit.cover)),
+              child: Container(
+                margin: EdgeInsets.only(left: 30, top: 10),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      'Tiêm chủng',
+                      style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 30,
+                          fontWeight: FontWeight.bold),
+                    ),
+                    Text(
+                      'Quản lí thời gian\n loại vaccin cho thú cưng',
+                      style: TextStyle(color: Colors.white, fontSize: 20),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+          GestureDetector(
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => CreateMealAppointmentScreen(),
+                ),
+              );
+            },
+            child: Container(
+              width: 500,
+              height: 130,
+              margin: EdgeInsets.all(20),
+              decoration: BoxDecoration(
+                  border: Border.all(color: Colors.white, width: 1),
+                  borderRadius: BorderRadius.circular(10),
+                  image: DecorationImage(
+                      image: AssetImage('assets/images/dinhduong.png'),
+                      fit: BoxFit.cover)),
+              child: Container(
+                margin: EdgeInsets.only(left: 30, top: 10),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      'Dinh dưỡng',
+                      style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 30,
+                          fontWeight: FontWeight.bold),
+                    ),
+                    Text(
+                      'Quản lí chế độ dinh dưỡng',
+                      style: TextStyle(color: Colors.white, fontSize: 20),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+          GestureDetector(
+            onTap: () {
+              Navigator.push(
+                context,
+                MaterialPageRoute(
+                  builder: (context) => CreateSizeAppointmentScreen(),
+                ),
+              );
+            },
+            child: Container(
+              width: 500,
+              height: 130,
+              margin: EdgeInsets.all(20),
+              decoration: BoxDecoration(
+                  border: Border.all(color: Colors.white, width: 1),
+                  borderRadius: BorderRadius.circular(10),
+                  image: DecorationImage(
+                      image: AssetImage('assets/images/dalieu.png'),
+                      fit: BoxFit.cover)),
+              child: Container(
+                margin: EdgeInsets.only(left: 30, top: 10),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.start,
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    Text(
+                      'Đo kích thước',
+                      style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 30,
+                          fontWeight: FontWeight.bold),
+                    ),
+                    Text(
+                      'Chăm sóc, spa cho thú cưng',
+                      style: TextStyle(color: Colors.white, fontSize: 20),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
